@@ -1,64 +1,68 @@
 {
-  "name": "aurora-launcher",
-  "version": "0.0.2",
-  "description": "Launcher for Minecraft",
-  "main": "build/main/index.js",
-  "scripts": {
-    "dev": "npm run build && npm run start:dev",
-    "build": "webpack",
-    "start:dev": "cross-env DEV=true electron .",
-    "start:prod": "electron .",
-    "clear": "rimraf build aurora-launcher-win32-x64",
-    "pack": "electron-packager build"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/AuroraTeam/Launcher.git"
-  },
-  "author": "AuroraTeam",
-  "contributors": [
-    "JoCat (https://github.com/JoCat)"
-  ],
-  "license": "GPL-3.0",
-  "bugs": {
-    "url": "https://github.com/AuroraTeam/Launcher/issues"
-  },
-  "homepage": "https://github.com/AuroraTeam/Launcher#readme",
-  "devDependencies": {
-    "@sweetalert2/theme-dark": "^4.0.3",
-    "@types/semver": "^7.3.4",
-    "@types/ws": "^7.4.0",
-    "aurora-api": "^0.1.3",
-    "copy-webpack-plugin": "^8.1.0",
-    "cross-env": "^7.0.3",
-    "css-loader": "^5.2.0",
-    "electron": "^13.6.6",
-    "electron-devtools-installer": "git+https://github.com/MarshallOfSound/electron-devtools-installer.git#9fad2da5a6961c80f83fc9d1ca601af59a1ae37a",
-    "electron-packager": "^15.2.0",
-    "file-loader": "^6.2.0",
-    "html-minimizer-webpack-plugin": "^2.1.0",
-    "less": "^4.1.1",
-    "less-loader": "^8.0.0",
-    "rimraf": "^3.0.2",
-    "sass": "^1.32.8",
-    "sass-loader": "^11.0.1",
-    "style-loader": "^2.0.0",
-    "stylus": "^0.54.8",
-    "stylus-loader": "^5.0.0",
-    "sweetalert2": "^10.15.7",
-    "ts-loader": "^8.0.18",
-    "typescript": "^4.2.3",
-    "vue": "^2.6.12",
-    "vue-loader": "^15.9.6",
-    "vue-router": "^3.5.1",
-    "vue-sweetalert2": "^4.2.1",
-    "vue-template-compiler": "^2.6.12",
-    "webpack": "^5.28.0",
-    "webpack-cli": "^4.5.0"
-  },
-  "dependencies": {
-    "p-map": "^4.0.0",
-    "semver": "^7.3.5",
-    "source-map-support": "^0.5.19"
-  }
+    "name": "aurora-launcher",
+    "version": "0.0.3",
+    "description": "Launcher for Minecraft",
+    "main": "build/main/index.js",
+    "private": true,
+    "scripts": {
+        "dev": "concurrently -n main,preload,renderer,electron -c blue.bold,yellow.bold,magenta.bold,cyan.bold \"npm:build:main -- --watch --sourcemap=inline\" \"npm:build:preload -- --watch --sourcemap=inline\" \"vite\" \"npm:start:dev\"",
+        "build": "concurrently -n main,preload,renderer -c blue.bold,yellow.bold,magenta.bold \"npm:build:main\" \"npm:build:preload\" \"npm:build:renderer\" --kill-others-on-fail",
+        "build:main": "esbuild src/main/index.ts --platform=node --bundle --outdir=build/main --external:electron --loader:.png=file",
+        "build:preload": "esbuild src/preload/index.ts --platform=node --bundle --outdir=build/preload --external:electron",
+        "build:renderer": "vite build",
+        "start:dev": "cross-env DEV=true nodemon --watch src --ext ts --ignore 'src/renderer' --exec electron .",
+        "start:prod": "electron .",
+        "clean": "rimraf build aurora-launcher-win32-x64",
+        "prettier": "prettier --config .prettierrc.json --write src",
+        "lint": "eslint src --ext .ts",
+        "lint:fix": "npm run lint -- --fix",
+        "pack": "electron-packager build"
+    },
+    "repository": {
+        "type": "git",
+        "url": "git+https://github.com/AuroraTeam/Launcher.git"
+    },
+    "author": "AuroraTeam",
+    "contributors": [
+        "JoCat (https://github.com/JoCat)"
+    ],
+    "license": "MIT",
+    "bugs": {
+        "url": "https://github.com/AuroraTeam/Launcher/issues"
+    },
+    "homepage": "https://github.com/AuroraTeam/Launcher#readme",
+    "devDependencies": {
+        "@sweetalert2/theme-dark": "^5.0.5",
+        "@types/semver": "^7.3.8",
+        "@types/ws": "^8.5.3",
+        "@typescript-eslint/eslint-plugin": "^5.18.0",
+        "@typescript-eslint/parser": "^5.18.0",
+        "aurora-api": "^0.2.0-beta.2",
+        "concurrently": "^7.1.0",
+        "cross-env": "^7.0.3",
+        "electron": "^19.0.3",
+        "electron-devtools-installer": "^3.2.0",
+        "electron-packager": "^15.4.0",
+        "esbuild": "^0.14.34",
+        "eslint": "^8.13.0",
+        "nodemon": "^2.0.15",
+        "prettier": "^2.4.1",
+        "rimraf": "^3.0.2",
+        "sass": "^1.47.0",
+        "sweetalert2": "11.4.8",
+        "typescript": "^4.5.4",
+        "vite": "^2.7.10",
+        "vite-plugin-vue2": "^2.0.1",
+        "vue": "^2.6.14",
+        "vue-router": "^3.5.3",
+        "vue-sweetalert2": "^5.0.2",
+        "vue-template-compiler": "^2.6.14"
+    },
+    "dependencies": {
+        "@auroralauncher/core": "^0.3.0",
+        "p-map": "^5.3.0",
+        "semver": "^7.3.5",
+        "skinview3d": "^2.2.0",
+        "source-map-support": "^0.5.21"
+    }
 }
