@@ -1,25 +1,105 @@
-import { AuroraAPI } from 'aurora-api'
-import { App } from '..'
-const apiConfig = require('@config').api
+import { AuroraAPI, Response, ResponseError } from 'aurora-api';
+import { api as apiConfig } from '@config';
+import { ipcMain } from 'electron';
 
-export default class APIManager {
-    api: AuroraAPI | null = null
+// TODO Подумать над реализацией корректной обработки запросов и отлова ошибок
+
+export class APIManager {
+    private readonly api = new AuroraAPI(
+        apiConfig.ws || 'ws://localhost:1370',
+        {
+            onOpen: () => (this.tryConnect = true),
+            onError: () => (this.tryConnect = true),
+        }
+    );
+    private tryConnect = false;
 
     constructor() {
-        this.connect()
+        ipcMain.handle('auth', (_, login: string, password: string) =>
+            this.auth(login, password)
+        );
+        ipcMain.handle('getStatus', () => this.getStatus());
+        ipcMain.handle('getServers', () => this.getServers());
+        ipcMain.handle('getProfile', (_, uuid: string) =>
+            this.getProfile(uuid)
+        );
     }
 
-    async connect() {
+    public getStatus(): 'connected' | 'failure' | 'connecting' {
+        if (this.api.hasConnected()) return 'connected';
+        if (this.tryConnect) return 'failure';
+        return 'connecting';
+    }
+
+    public async send(type: string, data?: object): Promise<Response> {
+        await this.api.ready();
+        return await this.api.send(type, data);
+    }
+
+    async auth(login: string, password: string): Promise<object> {
         try {
-            this.api = await new AuroraAPI().connect(apiConfig.url || 'ws://localhost:1370') as AuroraAPI
-            App.window.sendEvent('apiConnectSuccess')
+            const { data } = <Response>(
+                await this.send('auth', { login, password })
+            );
+            return data;
         } catch (error) {
-            App.window.sendEvent('apiConnectError', 'Ошибка при подключении')
-            console.log(error)
+            const e = <ResponseError>error; // Какого хрена TS?!
+            console.log(`Ошибка №${e.code}: ${e.message}`);
+            return {
+                error: e.message,
+            };
         }
     }
 
-    public async send(type: string, data?: object) {
-        return await this.api?.send(type, data)
+    async getServers(): Promise<any[] | object> {
+        try {
+            const { data } = <ServerResponse>await this.send('servers');
+            return data.servers;
+        } catch (error) {
+            const e = <ResponseError>error; // Какого хрена TS?!
+            return {
+                code: e.code,
+                message: e.message,
+            };
+        }
     }
+
+    async getProfile(uuid: string): Promise<any[] | object> {
+        try {
+            const { data } = <ProfileResponse>(
+                await this.send('profile', { uuid })
+            );
+            return data.profile;
+        } catch (error) {
+            const e = <ResponseError>error; // Какого хрена TS?!
+            return {
+                code: e.code,
+                message: e.message,
+            };
+        }
+    }
+
+    // TODO Доработать
+    public async getUpdates(dir: string): Promise<HashedFile[]> {
+        const { data } = <UpdatesResponse>await this.send('updates', { dir });
+        return data.hashes;
+    }
+}
+
+interface ServerResponse extends Response {
+    data: { servers: any[] };
+}
+
+interface ProfileResponse extends Response {
+    data: { profile: any[] };
+}
+
+interface UpdatesResponse extends Response {
+    data: { hashes: HashedFile[] };
+}
+
+interface HashedFile {
+    path: string;
+    hashsum: string;
+    size: number;
 }
