@@ -1,3 +1,4 @@
-import 'source-map-support/register'
-import Launcher from './core/Launcher'
-export const App = new Launcher()
+import 'source-map-support/register';
+
+import { Launcher } from './core/Launcher';
+new Launcher();
