@@ -1,63 +1,72 @@
 <template>
-    <div>
-        <div class="block">
-            <strong>Aurora Launcher</strong>
-            <p>Введите логин и пароль,<br> чтобы продолжить</p>
-            <form @submit.prevent="auth()">
-                <input type="text" placeholder="Логин" v-model="login">
-                <input type="password" placeholder="Пароль" v-model="password">
-                <button>Войти</button>
-            </form>
-        </div>
+    <div class="block">
+        <img class="logo" src="../assets/images/logo.png" />
+        <div>Aurora Launcher</div>
+        <p>
+            Введите логин и пароль,<br />
+            чтобы продолжить
+        </p>
+        <form @submit.prevent="auth()">
+            <input type="text" placeholder="Логин" v-model="login" />
+            <input type="password" placeholder="Пароль" v-model="password" />
+            <button>Войти</button>
+        </form>
     </div>
 </template>
 
 <script lang="ts">
-import Vue from 'vue'
-import LauncherAuth from '@scripts/LauncherAuth'
-import { Launcher } from '@Launcher'
+import Vue from 'vue';
 
 interface AuthResult {
-    error: string,
-    username: string,
-    userUUID: string
-    accessToken: string
+    error: string;
+    username: string;
+    userUUID: string;
+    accessToken: string;
 }
 
 export default Vue.extend({
     data() {
         return {
             login: '',
-            password: ''
-        }
+            password: '',
+        };
     },
     methods: {
         async auth() {
             // Валидацию можно делать как хошш))
-            if (this.login.length < 4) return this.showError('Логин должен быть не менее 4-ёх символов')
-            if (this.password.length < 8) return this.showError('Пароль должен быть не менее 8-ми символов')
+            if (this.login.length < 4)
+                return this.showError(
+                    'Логин должен быть не менее 4-ёх символов'
+                );
+            if (this.password.length < 8)
+                return this.showError(
+                    'Пароль должен быть не менее 8-ми символов'
+                );
 
-            const auth: AuthResult = await LauncherAuth.auth(this.login, this.password)
+            const auth: AuthResult = await launcherAPI.auth(
+                this.login,
+                this.password
+            );
             if (auth.error !== undefined) {
-                this.showError(auth.error)
+                this.showError(auth.error);
             } else {
-                Launcher.$emit('setUser', auth.username)
-                Launcher.$emit('showUser')
-                localStorage.setItem('username', auth.username)
-                localStorage.setItem('userUUID', auth.userUUID)
-                localStorage.setItem('accessToken', auth.accessToken)
-                this.$router.push('server-list')
+                this.$root.$emit('setUser', auth.username);
+                this.$root.$emit('showUser');
+                localStorage.setItem('username', auth.username); // @deprecated
+                localStorage.setItem('userUUID', auth.userUUID); // @deprecated
+                localStorage.setItem('accessToken', auth.accessToken); // @deprecated
+                this.$router.push('server-list');
             }
         },
         showError(message: string) {
             this.$swal({
                 title: 'Error!',
                 text: message,
-                icon: 'error'
-            })
-        }
-    }
-})
+                icon: 'error',
+            });
+        },
+    },
+});
 </script>
 
 <style lang="sass" scoped>
@@ -66,7 +75,7 @@ h1
 .block
     width: 300px
     height: 450px
-    background-color: #454BDF
+    background-color: #454BDFdf
     border-radius: 10px
     position: absolute
     top: 50px
@@ -75,14 +84,20 @@ h1
     display: flex
     flex-direction: column
     align-items: center
-    strong
+    img
+        height: 90px
+        margin-top: 35px
+        border-radius: 50%
+        filter: drop-shadow(0px 4px 4px rgba(0, 0, 0, 0.25))
+    div
         font-size: 24px
-        margin-top: 140px
+        margin-top: 20px
         text-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25)
     p
-        margin: 33px 0
-        font-size: 13px
-        line-height: 16px
+        margin: 25px 0 30px
+        font-size: 14px
+        line-height: 17px
+        text-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25)
         text-align: center
         font-family: 'Roboto', sans-serif
     form
@@ -101,8 +116,11 @@ h1
         font-family: 'Roboto', sans-serif
         outline: 0
         color: #fff
-    input::placeholder
-        color: #fff
+        &::placeholder
+            color: #fff
+        &:hover, &:focus
+            &::placeholder
+                color: #ddd
     button
         width: 194px
         height: 35px
@@ -114,4 +132,6 @@ h1
         outline: 0
         color: #fff
         box-shadow: 0px 0px 7px rgba(0, 0, 0, 0.25)
+        &:hover
+            box-shadow: 0px 0px 4px rgba(0, 0, 0, 0.25)
 </style>
